Still stuck. Remove body of test SendKeysAppendsText
If this will pass, it means that the culprit is in the body of this method. In this case we need to start adding back parts of the body

If this will fail, it means that the mere existence of this test method cause the problem. In this case we should ensure that nothing runs between tests, and if not, remove the bodies of other methods.