Build still passes. Adding back one more line.
If it will pass, this will be strange, as adding the last line will get us back to where we previously failed. In this case, add the last line just to make sure that it still fails, and rethink (maybe the problem was resolved by AppVeyor somehow?). Another thing to try is to add only the call to GetAttribute but without the Assert.

If it will fail, it will mean that the culprit is in the SendKeys, but it's still not clear why... try to simplify other tests to get a better understanding.